refactor: rename string capitalize function
